fix:replace HttpClient with HttpURLConnection for JDK 8 compatibility. (#1709)

+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-traffic-mirroring-plugin/src/main/java/com/tencent/cloud/plugin/trafficmirroring/config/TrafficMirroringProperties.java
@@ -40,7 +40,7 @@ public class TrafficMirroringProperties {
 	/**
 	 * Traffic mirroring request connection timeout in millisecond. Default is 5000.
 	 */
-	private long requestConnectionTimeout = 5000;
+	private int requestConnectionTimeout = 5000;
 
 	public boolean isEnabled() {
 		return enabled;
@@ -62,7 +62,7 @@ public class TrafficMirroringProperties {
 		return requestConnectionTimeout;
 	}
 
-	public void setRequestConnectionTimeout(long requestConnectionTimeout) {
+	public void setRequestConnectionTimeout(int requestConnectionTimeout) {
 		this.requestConnectionTimeout = requestConnectionTimeout;
 	}
